mindcare_api/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
bcrypt
python-multipart
email-validator
